requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
faster-whisper>=1.0.0
numpy>=1.26.0
soundfile>=0.12.1
scipy>=1.11.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.0.0
transformers>=4.46.0
sentencepiece>=0.2.0
TTS>=0.22.0
streamlit>=1.34.0
requests>=2.31.0
torch>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
